refactor: refactoring code and fixing some bugs
