+++ src/Pages/Packages/utils/packages.js
@@ -1,188 +1,92 @@
-const packageSchool = ({ age, withTransportation }) => {
-  const accomodation = [0];
-  let food = [280];
-  let transportation = withTransportation ? [160] : [0];
+import generatePackagesValues from './generatePackagesValues';
 
-  let foodDiscountDescription = '';
-  let transportationDiscountDescription = '';
+const XV_NOVEMBRO = 'Colégio XV de Novembro';
+const SEMINARIO = 'Seminário São José';
+const HOTEL_IBIS = 'Hotel Ibis';
 
-  //Alimentação
-  if (age <= 6) {
-    food = [280, 0];
-    foodDiscountDescription = 'Criança até 6 anos não paga alimentação';
-  } else if (age >= 7 && age <= 12) {
-    food = [280, 140];
-    foodDiscountDescription = 'Criança de 7 a 12 anos paga apenas 50% na alimentação';
-  }
+export const accommodations = [XV_NOVEMBRO, SEMINARIO, HOTEL_IBIS];
 
-  //Transporte
-  if (age <= 6) {
-    transportation = withTransportation ? [160, 0] : [0];
-    transportationDiscountDescription = 'Criança até 6 anos não paga transporte *(no colo dos pais)';
-  } else if (age >= 7 && age <= 12) {
-    transportation = withTransportation ? [160] : [0];
-  }
+const getPackages = (age) => {
+  const [schoollWithBuss, schollWithoutBuss] = generatePackagesValues('school', age);
+  const [seminaryWithBuss, seminaryWithoutBuss] = generatePackagesValues('seminary', age);
+  const [hotelWithBuss, hotelWithoutBuss] = generatePackagesValues('hotel', age);
 
-  const foodValue = food.length > 1 ? food[1] : food[0];
-  const transportationValue = transportation.length > 1 ? transportation[1] : transportation[0];
-
-  const total = accomodation[0] + foodValue + transportationValue;
-
-  return {
-    accomodation,
-    food,
-    transportation,
-    total,
-    discountDescription: {
-      food: foodDiscountDescription,
-      transportation: transportationDiscountDescription,
+  const packages = [
+    {
+      id: '1',
+      accomodation: { id: '1-colegio-individual', name: XV_NOVEMBRO },
+      title: 'PACOTE 1 - HOSPEDAGEM INDIVIDUAL EM SALA COLETIVA',
+      observation: '* Em salas de aula COM ônibus',
+      values: { ...schoollWithBuss },
+      transportation: 'Com Ônibus',
+      food: 'Café da manhã, almoço e jantar',
     },
-  };
-};
-
-const packageSeminary = ({ age, withTransportation }) => {
-  let accomodation = [600];
-  let food = [200];
-  let transportation = withTransportation ? [160] : [0];
-
-  let foodDiscountDescription = '';
-  let transportationDiscountDescription = '';
-  let accomodationDiscountDescription = '';
-
-  //Alimentação
-  if (age <= 6) {
-    food = [200, 0];
-    foodDiscountDescription = 'Criança até 6 anos não paga alimentação';
-  } else if (age >= 7 && age <= 12) {
-    food = [200, 100];
-    foodDiscountDescription = 'Criança de 7 a 12 anos paga apenas 50% na alimentação';
-  }
-
-  //Transporte
-  if (age <= 6) {
-    transportation = withTransportation ? [160, 0] : [0];
-    transportationDiscountDescription = 'Criança até 6 anos não paga transporte *(no colo dos pais)';
-  }
-
-  //Hospedagem
-  else if (age <= 8) {
-    accomodation = [600, 0];
-    accomodationDiscountDescription = 'Criança até 8 anos não paga hospedagem';
-  } else if (age >= 9 && age <= 14) {
-    accomodation = [600, 300];
-    accomodationDiscountDescription = 'Criança de 9 a 14 anos paga apenas 50% na hospedagem';
-  }
-
-  const accomodationValue = accomodation.length > 1 ? accomodation[1] : accomodation[0];
-  const foodValue = food.length > 1 ? food[1] : food[0];
-  const transportationValue = transportation.length > 1 ? transportation[1] : transportation[0];
-
-  const total = accomodationValue + foodValue + transportationValue;
-
-  return {
-    total,
-    accomodation,
-    food,
-    transportation,
-    discountDescription: {
-      food: foodDiscountDescription,
-      transportation: transportationDiscountDescription,
-      accomodation: accomodationDiscountDescription,
+    {
+      id: '2',
+      accomodation: { id: '2-colegio-individual', name: XV_NOVEMBRO },
+      title: 'PACOTE 2 - HOSPEDAGEM INDIVIDUAL EM SALA COLETIVA',
+      observation: '* Em salas de aula SEM ônibus',
+      values: { ...schollWithoutBuss },
+      transportation: 'Sem Ônibus',
+      food: 'Café da manhã, almoço e jantar',
     },
-  };
-};
-
-const packageHotel = ({ age, withTransportation }) => {
-  let accomodation = [550];
-  let food = [200];
-  let transportation = withTransportation ? [160] : [0];
-
-  let foodDiscountDescription = '';
-  let transportationDiscountDescription = '';
-  let accomodationDiscountDescription = '';
-
-  //Alimentação
-  if (age <= 6) {
-    food = [200, 0];
-    foodDiscountDescription = 'Criança até 6 anos não paga alimentação';
-  } else if (age >= 7 && age <= 12) {
-    food = [200, 100];
-    foodDiscountDescription = 'Criança de 7 a 12 anos paga apenas 50% na alimentação';
-  }
-
-  //Transporte
-  if (age <= 6) {
-    transportation = withTransportation ? [160, 0] : [0];
-    transportationDiscountDescription = 'Criança até 6 anos não paga transporte *(no colo dos pais)';
-  }
-
-  //Hospedagem
-  else if (age <= 10) {
-    accomodation = [550, 0];
-    accomodationDiscountDescription = 'Criança até 10 anos não paga hospedagem *(dormindo com os pais)';
-  }
-
-  const accomodationValue = accomodation.length > 1 ? accomodation[1] : accomodation[0];
-  const foodValue = food.length > 1 ? food[1] : food[0];
-  const transportationValue = transportation.length > 1 ? transportation[1] : transportation[0];
-
-  const total = accomodationValue + foodValue + transportationValue;
-
-  return {
-    total,
-    accomodation,
-    food,
-    transportation,
-    discountDescription: {
-      food: foodDiscountDescription,
-      transportation: transportationDiscountDescription,
-      accomodation: accomodationDiscountDescription,
+    {
+      id: '3',
+      accomodation: { id: '3-colegio-familia', name: XV_NOVEMBRO },
+      title: 'PACOTE 3 - HOSPEDAGEM FAMÍLIA EM SALA COLETIVA',
+      observation: '* Em salas de aula COM ônibus',
+      values: { ...schoollWithBuss },
+      transportation: 'Com Ônibus',
+      food: 'Café da manhã, almoço e jantar',
     },
-  };
+    {
+      id: '4',
+      accomodation: { id: '4-colegio-familia', name: XV_NOVEMBRO },
+      title: 'PACOTE 4 - HOSPEDAGEM FAMÍLIA EM SALA COLETIVA',
+      observation: '* Em salas de aula SEM ônibus',
+      values: { ...schollWithoutBuss },
+      transportation: 'Sem Ônibus',
+      food: 'Café da manhã, almoço e jantar',
+    },
+    {
+      id: '5',
+      accomodation: { id: '5-seminario-individual', name: SEMINARIO },
+      title: 'PACOTE 5 - HOSPEDAGEM INDIVIDUAL OU DUPLA',
+      observation: '* COM ônibus / Café da manhã incluso no seminário',
+      values: { ...seminaryWithBuss },
+      transportation: 'Com Ônibus',
+      food: 'Almoço e jantar',
+    },
+    {
+      id: '6',
+      accomodation: { id: '6-seminario-individual', name: SEMINARIO },
+      title: 'PACOTE 6 - HOSPEDAGEM INDIVIDUAL OU DUPLA',
+      observation: '* SEM ônibus / Café da manhã incluso no seminário',
+      values: { ...seminaryWithoutBuss },
+      transportation: 'Sem Ônibus',
+      food: 'Almoço e jantar',
+    },
+    {
+      id: '7',
+      accomodation: { id: '7-hotel-dupla', name: HOTEL_IBIS },
+      title: 'PACOTE 7 - HOSPEDAGEM DUPLA',
+      observation: '* COM ônibus / Café da manhã incluso no hotel',
+      values: { ...hotelWithBuss },
+      transportation: 'Com Ônibus',
+      food: 'Almoço e jantar',
+    },
+    {
+      id: '8',
+      accomodation: { id: '8-hotel-dupla', name: HOTEL_IBIS },
+      title: 'PACOTE 8 - HOSPEDAGEM DUPLA',
+      observation: '* SEM ônibus / Café da manhã incluso no hotel',
+      values: { ...hotelWithoutBuss },
+      transportation: 'Sem Ônibus',
+      food: 'Almoço e jantar',
+    },
+  ];
+
+  return packages;
 };
 
-const generatePackagesValues = (packageType, age) => {
-  if (packageType === 'school') {
-    const packageWithBuss = packageSchool({
-      age,
-      withTransportation: true,
-    });
-
-    const packageWithoutBuss = packageSchool({
-      age,
-      withTransportation: false,
-    });
-
-    return [packageWithBuss, packageWithoutBuss];
-  }
-
-  if (packageType === 'seminary') {
-    const packageWithBuss = packageSeminary({
-      age,
-      withTransportation: true,
-    });
-
-    const packageWithoutBuss = packageSeminary({
-      age,
-      withTransportation: false,
-    });
-
-    return [packageWithBuss, packageWithoutBuss];
-  }
-
-  if (packageType === 'hotel') {
-    const packageWithBuss = packageHotel({
-      age,
-      withTransportation: true,
-    });
-
-    const packageWithoutBuss = packageHotel({
-      age,
-      withTransportation: false,
-    });
-
-    return [packageWithBuss, packageWithoutBuss];
-  }
-};
-
-export default generatePackagesValues;
+export default getPackages;
